# swagger.yaml
swagger: "2.0"
info:
  description: "Used to get details about codes and code lists stored by ONS.
  Codes are used to provide a common definition when presenting statistics
  with related categories. Codes are gathered in code lists, which may change over time
  to include new or different codes. The meaning of a code should not change over
  time, but new codes may be created where new meaning is required."
  version: "1.0.0"
  title: "Explore codes and lists"
  license:
    name: "Open Government Licence v3.0"
    url: "http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/"
basePath: "/v1"
schemes:
- "http"
parameters:
  id:
    name: id
    description: "The ID of a code list"
    in: path
    type: string
    required: true
  edition:
    name: edition
    description: "The ID of a code list edition"
    in: path
    type: string
    required: true
  codeId:
    name: code_id
    description: "The ID of the Code within an edition of a code list"
    in: path
    type: string
    required: true
paths:
  /code-lists:
    get:
      tags:
      - "Code Lists"
      summary: "Get a list of code lists"
      description: "Get a set of code lists containing information about dimensions which are used for all datasets at the ONS"
      produces:
      - "application/json"
      responses:
        200:
          description: "A Json message containing a set of code lists"
          schema:
            $ref: '#/definitions/CodeLists'
        404:
          description: "CodeLists not found"
        500:
          description: "Failed to process the request due to an internal error"
  /code-lists/{id}:
    get:
      tags:
      - "Code Lists"
      summary: "Get a code list"
      description: "Get information about a code list"
      parameters:
      - $ref: '#/parameters/id'
      produces:
      - "application/json"
      responses:
        200:
          description: "Json object containing information about the code list"
          schema:
            $ref: '#/definitions/CodeList'
        404:
          description: "Code list not found"
        500:
          description: "Failed to process the request due to an internal error"
  /code-lists/{id}/editions:
    get:
      tags:
      - "Editions"
      summary: "Get a list of editions"
      description: "Get a list of editions for a code list"
      parameters:
      - $ref: '#/parameters/id'
      produces:
      - "application/json"
      responses:
        200:
          description: "Json object containing information about the code list editions"
          schema:
            $ref: '#/definitions/Editions'
        404:
          description: "Editions for code list not found"
        500:
          description: "Failed to process the request due to an internal error"
  /code-lists/{id}/editions/{edition}:
    get:
      tags:
      - "Editions"
      summary: "Get an edition of a code list"
      description: "Get information about an edition of a code list"
      parameters:
      - $ref: '#/parameters/id'
      - $ref: '#/parameters/edition'
      produces:
      - "application/json"
      responses:
        200:
          description: "Json object containing information about the code list edition"
          schema:
            $ref: '#/definitions/Edition'
        404:
          description: "Edition for code list not found"
        500:
          description: "Failed to process the request due to an internal error"
  /code-lists/{id}/editions/{edition}/codes:
    get:
      tags:
       - "Codes"
      summary: "Get a list of all codes within a code list edition"
      description: "Get a list of codes within a code list edition"
      parameters:
      - $ref: '#/parameters/id'
      - $ref: '#/parameters/edition'
      produces:
      - "application/json"
      responses:
        200:
          description: "A Json message containing a list of Codes"
          schema:
            $ref: '#/definitions/Codes'
        404:
          description: "Edition not found"
        500:
          description: "Failed to process the request due to an internal error"
  /code-lists/{id}/editions/{edition}/codes/{code_id}:
    get:
      tags:
       - "Codes"
      summary: "Get a code from a edition of a code list"
      description: "Get information about a code within a code list"
      parameters:
      - $ref: '#/parameters/id'
      - $ref: '#/parameters/edition'
      - $ref: '#/parameters/codeId'
      produces:
      - "application/json"
      responses:
        200:
          description: "Get in depth information about a code"
          schema:
            $ref: '#/definitions/Code'
        404:
          description: "Code list or Edition or code not found"
        500:
          description: "Failed to process the request due to an internal error"
  /code-lists/{id}/editions/{edition}/codes/{code_id}/datasets:
    get:
      tags:
       - "Datasets"
      summary: "Get datasets"
      description: "Get a list of the datasets that use this code"
      parameters:
      - $ref: '#/parameters/id'
      - $ref: '#/parameters/edition'
      - $ref: '#/parameters/codeId'
      produces:
      - "application/json"
      responses:
        200:
          description: "Get a list of the datasets that use this code"
          schema:
            $ref: '#/definitions/Datasets'
        404:
          description: "Code not found"
        500:
          description: "Failed to process the request due to an internal error"
definitions:
  CodeList:
    type: object
    properties:
      links:
        type: object
        properties:
          self:
            $ref: '#/definitions/SelfHref'
          editions:
            $ref: '#/definitions/Href'
  CodeLists:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: '#/definitions/CodeList'
      number_of_results:
        type: integer
        description: "The number of results available"

  Edition:
    type: object
    properties:
      edition:
        type: string
        description: "The edition"
      label:
        type: string
        description: "A label used by the edition"
      links:
        type: object
        properties:
          self:
            $ref: '#/definitions/Href'
          codes:
            $ref: '#/definitions/Href'
          editions:
            $ref: '#/definitions/Href'
  Editions:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: '#/definitions/Edition'
      count:
        $ref: '#/definitions/Count'
      total_count:
        $ref: '#/definitions/TotalCount'
      limit:
        $ref: '#/definitions/Limit'
  Code:
    type: object
    properties:
      id:
        type: string
        description: "The value of a code"
      label:
        type: string
        description: "A label used by the code"
      links:
        type: object
        properties:
          self:
            $ref: '#/definitions/Href'
          code_list:
            $ref: '#/definitions/Href'
          datasets:
            $ref: '#/definitions/Href'
  Codes:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: '#/definitions/Code'
      count:
        $ref: '#/definitions/Count'
      total_count:
        $ref: '#/definitions/TotalCount'
      limit:
        $ref: '#/definitions/Limit'
  Dataset:
    type: object
    properties:
      links:
        type: object
        properties:
          self:
            $ref: '#/definitions/Href'
      dimension_label:
        type: string
        description: "The code label used by this dataset"
      editions:
        items:
          type: array
          items:
            $ref: '#/definitions/DatasetEdition'
  DatasetEdition:
    type: object
    properties:
      links:
        type: object
        properties:
          self:
            $ref: '#/definitions/Href'
          dataset_dimension:
            $ref: '#/definitions/Href'
          latest_version:
            $ref: '#/definitions/Href'
  Datasets:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: '#/definitions/Dataset'
      count:
        $ref: '#/definitions/Count'
      total_count:
        $ref: '#/definitions/TotalCount'
      limit:
        $ref: '#/definitions/Limit'
  Href:
    type: object
    properties:
      id:
        type: string
        description: "The id of the resource"
      href:
        type: string
        description: "The URL to the resource"
  Count:
    type: string
    description: "The number of results returned in this response"
  TotalCount:
    type: string
    description: "The total number of results available"
  Limit:
    type: string
    description: "The limit applied to the number of results returned"